Organize stats modules
