+++ src/js/stats/statsRender.js
